Simpler API and made it immutable by default
All package level functions regarding vector arithmetic has been
removed. Instead methods should be called on a vector which now are
immutable by default. The concept of an UnsafeVector as been introduced
it follows the same API as a normal Vector, the difference is that an
UnsafeVector does all arithmetic operations in-place and shall therefore
be used with care, hence the name UnsafeVector.